optimize modal style and copy event
