+++ src/assets/styles/lib.scss
@@ -64,6 +64,7 @@
 .mb-15{margin-bottom:15px}
 .mb-20{margin-bottom:20px}
 .pt-10{padding-top: 10px}
+.pl-15{padding-left: 15px}
 .link-hover{
   transition: color .3s;
 }
