+++ src/assets/styles/style.scss
@@ -98,3 +98,10 @@
       transform: rotate(1turn);
   }
 }
+
+.oa {
+  overflow: hidden;
+}
+.oa:hover{
+  overflow: auto;
+}
