+++ src/views/components/trace/trace-detail-chart-list.vue
@@ -25,7 +25,7 @@
         <span>{{i}}</span>
       </span>
     </transition-group>
-    <rk-sidebox :show.sync="showDetail" :title="$t('spanInfo')">
+    <rk-sidebox :width="'50%'" :show.sync="showDetail" :title="$t('spanInfo')">
       <div class="rk-trace-detail">
         <h5 class="mb-15">{{$t('tags')}}.</h5>
         <div class="mb-10 clear"><span class="g-sm-4 grey">{{$t('endpoint')}}:</span><span class="g-sm-8 wba">{{this.currentSpan.label}}</span></div>
@@ -33,12 +33,20 @@
         <div class="mb-10 clear"><span class="g-sm-4 grey">{{$t('component')}}:</span><span class="g-sm-8 wba">{{this.currentSpan.component}}</span></div>
         <div class="mb-10 clear"><span class="g-sm-4 grey">Peer:</span><span class="g-sm-8 wba">{{this.currentSpan.peer||'No Peer'}}</span></div>
         <div class="mb-10 clear"><span class="g-sm-4 grey">{{$t('error')}}:</span><span class="g-sm-8 wba">{{this.currentSpan.isError}}</span></div>
-        <div class="mb-10 clear" v-for="i in this.currentSpan.tags" :key="i.key"><span class="g-sm-4 grey">{{i.key}}:</span><span class="g-sm-8 wba">{{i.value}}</span></div>
+        <div class="mb-10 clear" v-for="i in this.currentSpan.tags" :key="i.key">
+          <span class="g-sm-4 grey">{{i.key}}:</span>
+          <span class="g-sm-8 wba">
+            {{i.value}}
+            <svg v-if="i.key==='db.statement'" class="icon vm grey link-hover cp ml-5" @click="copy(i.value)">
+              <use xlink:href="#review-list"></use>
+            </svg>
+          </span>
+        </div>
         <h5 class="mb-10" v-if="this.currentSpan.logs" v-show="this.currentSpan.logs.length">{{$t('logs')}}.</h5>
         <div v-for="(i, index) in this.currentSpan.logs" :key="index">
           <div class="mb-10 sm"><span class="mr-10">{{$t('time')}}:</span><span class="grey">{{i.time | dateformat}}</span></div>
           <div class="mb-15 clear" v-for="(_i, _index) in i.data" :key="_index">
-            <div class="mb-10">{{_i.key}}:</div><pre class="g-sm-8 mt-0 mb-0 sm" style="overflow:auto">{{_i.value}}</pre>
+            <div class="mb-10">{{_i.key}}:</div><pre class="pl-15 mt-0 mb-0 sm oa" >{{_i.value}}</pre>
           </div>
         </div>
       </div>
@@ -49,6 +57,7 @@
   </div>
 </template>
 <script lang="js">
+import copy from '@/utils/copy'
 import * as d3 from 'd3';
 import Trace from './d3-trace';
 /* eslint-disable */
@@ -82,6 +91,7 @@ export default {
     // this.computedScale();
   },
   methods: {
+    copy,
     handleSelectSpan(i) {
       this.currentSpan = i.data;
       this.showDetail = true;
